feat: switch reward points to BigDecimal and update DTOs
- Preserve fractional reward points (e.g., 91.5 instead of 91)
- Updated RewardPointsService and DTOs to use BigDecimal
- Added unit tests for multiple customers and months
- Reformatted codebase to consistent 4-space indentation
- Moved sample JSON files into docs folder

+++ src/main/java/com/rewardpointsapi/dto/CustomerRewardPointsSummaryDTO.java
@@ -1,13 +1,14 @@
 package com.rewardpointsapi.dto;
 
+import java.math.BigDecimal;
 import java.util.Map;
 
 public class CustomerRewardPointsSummaryDTO {
 
     private String customerId;
     private String customerName;
-    private Map<String, Integer> monthlyRewards;
-    private int totalRewards;
+    private Map<String, BigDecimal> monthlyRewards;
+    private BigDecimal totalRewards;
 
     // Empty constructor
     public CustomerRewardPointsSummaryDTO() {
@@ -15,7 +16,7 @@ public class CustomerRewardPointsSummaryDTO {
 
     // Parameterized constructor
     public CustomerRewardPointsSummaryDTO(String customerId, String customerName, 
-                                          Map<String, Integer> monthlyRewards, int totalRewards) {
+                                          Map<String, BigDecimal> monthlyRewards, BigDecimal totalRewards) {
         this.customerId = customerId;
         this.customerName = customerName;
         this.monthlyRewards = monthlyRewards;
@@ -39,19 +40,19 @@ public class CustomerRewardPointsSummaryDTO {
         this.customerName = customerName;
     }
 
-    public Map<String, Integer> getMonthlyRewards() {
+    public Map<String, BigDecimal> getMonthlyRewards() {
         return monthlyRewards;
     }
 
-    public void setMonthlyRewards(Map<String, Integer> monthlyRewards) {
+    public void setMonthlyRewards(Map<String, BigDecimal> monthlyRewards) {
         this.monthlyRewards = monthlyRewards;
     }
 
-    public int getTotalRewards() {
+    public BigDecimal getTotalRewards() {
         return totalRewards;
     }
 
-    public void setTotalRewards(int totalRewards) {
+    public void setTotalRewards(BigDecimal totalRewards) {
         this.totalRewards = totalRewards;
     }
 }
